perform key package validation on (initial) fetch and updates of the app configuration
