+++ src/xcode/ENA/ENA/Source/Services/DownloadedPackagesStore/__tests__/Doubles/DownloadedPackagesStoreErrorStub.swift
@@ -76,4 +76,6 @@ class DownloadedPackagesStoreErrorStub: DownloadedPackagesStoreV2 {
 
 	var keyValueStore: Store?
 
+	var revokationList: [String] = []
+
 }
